dump.sql and docker

+++ dump.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users(
     user_id serial PRIMARY KEY,
     username VARCHAR (64) UNIQUE NOT NULL,
-    created_at INTEGER (128) NOT NULL
+    created_at bigint NOT NULL
 );
